+++ ryoku/shell/quickshell/shell/modules/bar/barstyles/qsbar/BarSlot.qml
@@ -9,6 +9,7 @@ import Quickshell
 import Quickshell.Wayland
 import Quickshell.Io
 import "modules"
+import shell.services as Svc
 
 PanelWindow {
     id: barSlot
@@ -391,9 +392,8 @@ PanelWindow {
                         anchors.verticalCenter: parent.verticalCenter
                         text: {
                             clock.now;
-                            var days = ["Sun","Mon","Tue","Wed","Thu","Fri","Sat"];
                             var d = new Date();
-                            return days[d.getDay()] + " " + d.getDate();
+                            return Svc.Config.formatLoc.standaloneDayName(d.getDay(), 1) + " " + d.getDate();
                         }
                         color: Qt.rgba(barSlot.root.ink.r, barSlot.root.ink.g, barSlot.root.ink.b, 0.5)
                         font.family: barSlot.root.mono